requirements.txt
google-genai
kagglehub
pandas
pysbd
python-dotenv
